+++ src/main/java/net/silentchaos512/gear/data/client/ModBlockStateProvider.java
@@ -93,14 +93,6 @@ public class ModBlockStateProvider extends BlockStateProvider {
                             .texture("particle", front));
         }
 
-        getVariantBuilder(ModBlocks.METAL_PRESS.get()).forAllStates(state -> {
-            return ConfiguredModel.builder()
-                    .modelFile(models().cubeTop("metal_press",
-                            modLoc("block/metal_press_top"),
-                            modLoc("block/metal_press_side")))
-                    .build();
-        });
-
         // Compounders
         {
             BlockModelBuilder offModel = models().orientableWithBottom("metal_alloyer",